improved deployment script

+++ deployment/run_container.sh
@@ -14,11 +14,22 @@ name=$1
 port=$2
 env=$3
 
+try_stop_and_remove_container () {
+	if [ "$(docker ps -f name="$name" -q)" ]; then
+		echo trying to stop "$name" container &&
+		docker stop "$name" &&
+		echo stopped "$name" container
+	fi
+
+	if [ "$(docker ps -f name="$name" -q -a)" ]; then
+		echo trying to remove "$name" container &&
+		docker container rm "$name" &&
+		echo removed "$name" container
+	fi
+}
+
 docker build -t "$name" -f Docker/build.dockerfile . &&
-docker stop "$name" &&
-echo stopped "$name" container &&
-docker container rm "$name" &&
-echo removed "$name" container &&
+try_stop_and_remove_container &&
 docker run \
 	-e ASPNETCORE_ENVIRONMENT="$env" \
 	-d \
@@ -26,5 +37,6 @@ docker run \
 	-t \
 	--restart=unless-stopped \
 	-v "$HOME"/Logs/"$name":/app/Serilogs \
+	--name "$name" \
 	"$name" &&
 echo looks like container "$name" should start
